+++ memiavl/snapshot_test.go
@@ -165,7 +165,7 @@ func TestDBSnapshotRestore(t *testing.T) {
 
 func testSnapshotRoundTrip(t *testing.T, db *DB) {
 	t.Helper()
-	exporter, err := NewMultiTreeExporter(db.dir, uint32(db.Version()), true)
+	exporter, err := NewMultiTreeExporter(db.dir, uint64(db.Version()), true)
 	require.NoError(t, err)
 
 	restoreDir := t.TempDir()
